# apps/cloud-api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TelaFlow Cloud API",
    "description": "API da TelaFlow Cloud \u2014 fase inicial: governan\u00e7a e integra\u00e7\u00e3o, sem persist\u00eancia completa.",
    "version": "0.1.0"
  },
  "paths": {
    "/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "List Events",
        "description": "Lista eventos em mem\u00f3ria (ordem de cria\u00e7\u00e3o).",
        "operationId": "list_events_events_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Events Events Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Create Event",
        "description": "Cria evento em mem\u00f3ria. N\u00e3o \u00e9 CRUD completo \u2014 apenas stub para integra\u00e7\u00e3o inicial.",
        "operationId": "create_event_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Event Events Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Event",
        "description": "Retorna um evento por id ou 404.",
        "operationId": "get_event_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Event Events  Event Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/events/{event_id}/draw-configs": {
      "get": {
        "tags": [
          "draw-configs"
        ],
        "summary": "List Draw Configs",
        "operationId": "list_draw_configs_events__event_id__draw_configs_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Draw Configs Events  Event Id  Draw Configs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "draw-configs"
        ],
        "summary": "Create Draw Config",
        "operationId": "create_draw_config_events__event_id__draw_configs_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DrawConfigCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Draw Config Events  Event Id  Draw Configs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}/draw-configs/{draw_config_id}": {
      "get": {
        "tags": [
          "draw-configs"
        ],
        "summary": "Get Draw Config",
        "operationId": "get_draw_config_events__event_id__draw_configs__draw_config_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "draw_config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draw Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Draw Config Events  Event Id  Draw Configs  Draw Config Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "draw-configs"
        ],
        "summary": "Update Draw Config",
        "operationId": "update_draw_config_events__event_id__draw_configs__draw_config_id__patch",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "draw_config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draw Config Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DrawConfigUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Draw Config Events  Event Id  Draw Configs  Draw Config Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "draw-configs"
        ],
        "summary": "Delete Draw Config",
        "operationId": "delete_draw_config_events__event_id__draw_configs__draw_config_id__delete",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "draw_config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draw Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Draw Config Events  Event Id  Draw Configs  Draw Config Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/events/{event_id}/export-readiness": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Readiness",
        "description": "Checagens m\u00ednimas server-side para futuro ExportPackage.\nN\u00e3o gera pack nem assinatura \u2014 s\u00f3 estrutura e consist\u00eancia declar\u00e1vel.",
        "operationId": "export_readiness_events__event_id__export_readiness_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Export Readiness Events  Event Id  Export Readiness Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}/export": {
      "post": {
        "tags": [
          "export"
        ],
        "summary": "Run Export",
        "description": "Gera snapshot JSON do evento (artefatos l\u00f3gicos). Exige `export_readiness.ready == true`.\nN\u00e3o gera ZIP nem assinatura (MVP incremental).",
        "operationId": "run_export_events__event_id__export_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Run Export Events  Event Id  Export Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DrawConfigCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Name"
          },
          "max_winners": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Max Winners",
            "default": 1
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "DrawConfigCreate",
        "description": "Cria\u00e7\u00e3o \u2014 `draw_config_id` e `event_id` definidos pela rota."
      },
      "DrawConfigUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "max_winners": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 999.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Winners"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DrawConfigUpdate",
        "description": "Atualiza\u00e7\u00e3o parcial (PATCH)."
      },
      "Event": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 12,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Event Id"
          },
          "organization_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 12,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Organization Id"
          },
          "name": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "organization_id",
          "name"
        ],
        "title": "Event",
        "description": "Ids opacos \u2014 alinhado a shared-contracts (charset e comprimento; formato final na persist\u00eancia)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}